% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_climate}
\alias{d_climate}
\title{Climate input}
\format{
A data frame with 156 rows and 14 variables:
\describe{
  \item{year}{calendar year}
  \item{month}{month}
  \item{tmp_min}{monthly mean daily minimum temperature (C)}
  \item{tmp_max}{monthly mean daily maximum temperature (C)}
  \item{tmp_ave}{monthly mean daily average temperature (C). (optional)}
  \item{prcp}{monthly rainfall (mm month-1)}
  \item{rh}{monthly average relative humidity (\%)}.
  \item{srad}{monthly mean daily solar radiation (MJ m-2 d-1)}
  \item{frost_days}{frost days per month (d month-1)}
  \item{co2}{monthly mean atmospheric co2 (ppm), required if calculate_d13c=1 (optional)}
  \item{d13catm}{Monthly mean isotopic composition of air (‰), required if calculate_d13c=1 (optional)}
  \item{n_depo}{monthly nitrogen deposition (tN/ha) (optional)}
  \item{weibull_a}{wind speed weibull a parameter, required if wind_dist=1 (optional)}
  \item{weibull_k}{wind speed weibull k parameter, required if wind_dist=1 (optional)}
}
}
\usage{
d_climate
}
\description{
Table containing the information about monthly values for climatic data.
}
\keyword{datasets}
