% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_site}
\alias{d_site}
\title{Site input}
\format{
A data frame with 1 rows and 14 variables:
\describe{
  \item{latitude}{site latitude in the WGS84 coordinate system}
  \item{altitude}{site altitude, m a.s.l.}
  \item{soil_class}{ soil class, according to table 2 user manual of 3PGpjs. 1 - Sandy; 2 - Sandy loam; 3 - Clay loam; 4 - Clay; 0 - No effect of available soil water on production}
  \item{asw_i}{initial available soil water (mm)}
  \item{asw_max}{minimum available soil water (mm)}
  \item{asw_min}{maximum available soil water (mm)}
  \item{from}{year and month indicating the start of simulation. Provided in form of year-month. E.g. "2000-01"}
  \item{to}{year and month indicating the end of simulation. Provided in form of year-month. E.g. "2009-12", will include December 2009 as last simulation month}
  \item{soil_carbon}{soil carbon stock (tC/ha)}
  \item{soil_nitrogen}{soil nitrogen stock (tN/ha)}
  \item{soil_phosphorus}{soil phosphorus stock (tP/ha)}
  \item{bb_cbp}{spatial scaling for bark beetle infestation}
  \item{bb_cst}{temporal scaling for bark beetle infestation}
  \item{background_infestation}{bark beetle background infestation (default = 1)}
}
}
\usage{
d_site
}
\description{
Table containing the information about site conditions.
}
\keyword{datasets}
