% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{prepare_input}
\alias{prepare_input}
\title{Check and prepare input for running 3-PG model}
\usage{
prepare_input(site, species, climate, thinning = NULL, parameters = NULL,
  size_dist = NULL, settings = NULL)
}
\arguments{
\item{site}{table containing the information about site conditions.
\itemize{
\item latitude: site latitude in the WGS84 coordinate system.
\item altitude: site altitude, m a.s.l.
\item soil_class: 1 - Sandy; 2 - Sandy loam; 3 - Clay loam; 4 - Clay; 0 - No effect of asw on production.
\item asw_i: initial available soil water (mm).
\item asw_min: minimum available soil water (mm).
\item asw_max: maximum available soil water (mm).
\item from: year and month indicating the start of simulation. Provided in form of year-month. E.g. "2000-01".
\item to: year and month indicating the end of simulation. Provided in form of year-month. E.g. "2009-12", will include December 2009 as last simulation month
\item soil_carbon: soil carbon stock (tC/ha)
\item soil_nitrogen: soil nitrogen stock (tN/ha)
\item soil_phosphorus: soil phosphorus stock (tP/ha)
\item bb_cbp: spatial scaling for bark beetle infestation
\item bb_cst: temporal scaling for bark beetle infestation
\item background_infestation: bark beetle background infestation (default = 1)
}}

\item{species}{table containing the information about species level data. Each row corresponds to one species/cohort.
\itemize{
\item species: species or cohort id/name. It must be consistent with species names in \code{thinning}, \code{parameters} and \code{sizeDist} tables.
\item planted: year and month indicating when species was planted. Provided in form of year-month. E.g. "2000-01".
\item fertility: soil fertility for a given species. Range from 0 to 1.
\item stems_n: number of trees per ha.
\item biom_stem: stem biomass for a given species (Mg/ha).
\item biom_root: root biomass for a given species (Mg/ha).
\item biom_foliage: initial foliage biomass (Mg/ha). If this is a leafless period, provide the spring foliage biomass.
\item deadwood_carbon: carbon stock in deadwood (Mg/ha)
\item deadwood_nitrogen: nitrogen stock in deadwood (Mg/ha)
\item litter_carbon: carbon stock in litter (Mg/ha)
\item litter_nitrogen: nitrogen stock in litter (Mg/ha)
\item n_sapling: number of saplings in regenerated stand (N/ha)
\item sapling_wf: foliage biomass of saplings (Mg/ha)
\item sapling_wr: root biomass of saplings (Mg/ha)
\item sapling_ws: stem biomass of saplings (Mg/ha)
\item cwd_removal: share of removed deadwood
\item salvage_wind: share of salvaged stock after wind disturbances
\item salvage_biotic: share of salvaged stock after biotic disturbances
\item salvage_fire: share of salvaged stock after wildfires
}}

\item{climate}{table containing the information about monthly values for climatic data. If the climate table has exactly 12 rows it will be replicated for the number of years and months specified by \code{from} - \code{to}. Otherwise, it will be subsetted to the selected time period. More details about preparing climate data are at \code{\link{prepare_climate}}.
\itemize{
\item year: year of observation (only required for subsetting) (optional).
\item month: months of observation (only required for subsetting) (optional).
\item tmp_min: monthly mean daily minimum temperature (C).
\item tmp_max: monthly mean daily maximum temperature (C).
\item tmp_ave: monthly mean daily average temperature (C) (optional).
\item prcp: monthly rainfall (mm month-1).
\item rh: monthly average relative humidity (%).
\item srad: monthly mean daily solar radiation (MJ m-2 d-1).
\item frost_days: frost days per month (d month-1).
\item vpd_day: frost days per month (mbar) (optional).
\item co2: monthly mean atmospheric co2 (ppm), required if calculate_d13c=1 (optional)
\item d13catm: monthly mean isotopic composition of air (‰), required if calculate_d13c=1 (optional)
\item n_depo: monthly nitrogen deposition (tN/ha) (optional)
\item weibull_a: wind speed weibull a parameter, required if wind_dist=1 (optional)
\item weibull_k: wind speed weibull k parameter, required if wind_dist=1 (optional)
}}

\item{thinning}{table containing the information about thinnings. If there is no thinning, it must be \code{NULL}.
\itemize{
\item species: species or cohort id/name. It must be consistent with species names in \code{species}, \code{parameters} and \code{sizeDist} tables.
\item age: age when thinning is performed.
\item stems_n: number of trees remaining after thinning
\item foliage: type of thinning (above/below). Default is 1.
\item root: type of thinning (above/below). Default is 1.
\item stem: type of thinning (above/below). Default is 1.
\item BA: share of basal area removal or target basal area after thinning.
\item Vol: share of volume removal or target volume after thinning.
\item crop_trees: number of crop trees per hectare.
}}

\item{parameters}{table containing the information about parameters to be modified. Values that are not provided are replaced by defaults.
\itemize{
\item parameter: name of the parameter, must be consistent in naming with \code{\link{i_parameters}}
\item species: each column must correspond to species/cohort id/name, as defined in \code{species} table
}}

\item{size_dist}{table containing the information about size distribution to be modified. Values that are not provided are replaced by defaults.
\itemize{
\item parameter: name of the parameter, must be consistent in naming with \code{\link{i_sizeDist}}
\item species: each column must correspond to species/cohort id/name, as defined in \code{species} table
}}

\item{settings}{a list with settings for the model. Values that are not provided are replaced by defaults.
\itemize{
\item light_model: `1` - 3-PGpjs (default); `2` - 3-PGmix
\item transp_model: `1` - 3-PGpjs (default); `2` - 3-PGmix
\item phys_model:  `1` - 3-PGpjs (default); `2` - 3-PGmix
\item height_model: `1` - linear (default); `2` - non-linear
\item correct_bias: `0` - no (default); `1` - yes
\item calculate_d13c: `0` - no (default); `1` - yes
\item soil_model: `0` - no soil model (default); `1` - ICBM/2N; `2` - Yasso
\item gpp_model: `1` - 3-PGpjs (default); `2` - P-model; `3` - P-hydro model
\item method_jmaxlim:  `1` - Wang; `2` - Smith; `3` - none (default)
\item nut_lim: `0` - fixed fr (default); `2` - dynamic fr based on N/P demand and soil available N/P
\item maint_resp: `1` - fixed NPP/GPP ratio (default); `2` - dynamic maintenance respiration
\item canopy_cond: `1` - Jarvis (default); `2` - P-model; `3` - P-hydro model
\item mortality_model: `1` - 3-PGpjs (default); `2` - Brandl S(t); `3` - LPJ-GUESS; `4` - iLand
\item wind_dist: `0` - no wind disturbance (default); `1` - ForestGALES TMC
\item beetle_dist:  `0` no bark beetle disturbance (default); `2` - LandClim bark beetle disturbance
\item d_sus_type :  `1` bark beetle vitality effect based on f_sw (default); `2` - bark beetle vitality effect based on growth efficiency
\item dist_start :  month when disturbance are initially activated
\item spruce_idx :  index of spruce in the species list
}}
}
\value{
a list with seven tables. Each table corresponds to one of the input tables.
}
\description{
Checks and prepares all input tables to be used in \code{\link{run_3PG}}. For detailed descriptions see Forrester (2020).
}
\details{
This function checks and prepares the input data for the \code{\link{run_3PG}}. The output is a list with 7 tables. Each of them corresponds to the one from input.
}
\examples{
prepare_input( site = d_site, species = d_species, climate = d_climate, d_thinning)
}
\references{
Forrester, D. I., 2020. 3-PG User Manual. Swiss Federal Institute for Forest, Snow and Landscape Research WSL, Birmensdorf, Switzerland. 70 p. Available at the following web site: \url{http://sites.google.com/site/davidforresterssite/home/projects/3PGmix/3pgmixdownload}

Sands, P. J., 2010. 3PGpjs user manual. Available at the following web site: \url{http://3pg.sites.olt.ubc.ca/files/2014/04/3PGpjs_UserManual.pdf}
}
\seealso{
\code{\link{run_3PG}}, \code{\link{prepare_parameters}}, \code{\link{prepare_sizeDist}}, \code{\link{prepare_thinning}}, \code{\link{prepare_climate}}
}
